+++ adapter/runtime/pom.xml
@@ -28,7 +28,7 @@
         <maven.compiler.target>8</maven.compiler.target>
         <commons.io.version>2.8.0</commons.io.version>
         <protobuf.version>3.8.0</protobuf.version>
-        <grpc.version>1.15.0</grpc.version>
+        <grpc.version>1.50.0</grpc.version>
         <junit.version>4.10</junit.version>
         <fasterxml.version>2.10.0</fasterxml.version>
     </properties>
