+++ adapter/runtime/src/main/java/org/apache/rocketmq/eventbridge/adapter/runtime/pipline/callback/TaskEventCallBack.java
@@ -1,0 +1,14 @@
+package org.apache.rocketmq.eventbridge.adapter.runtime.pipline.callback;
+
+public class TaskEventCallBack {
+
+    private ICallbackFunction callbackFunction;
+
+    public void addCallBack(ICallbackFunction callbackFunction) {
+        this.callbackFunction = callbackFunction;
+    }
+
+    public ICallbackFunction getCallbackFunction() {
+        return callbackFunction;
+    }
+}
