+++ adapter/runtime/src/main/java/org/apache/rocketmq/eventbridge/adapter/runtime/scheduler/Pipeline.java
@@ -2,7 +2,7 @@ package org.apache.rocketmq.eventbridge.adapter.runtime.scheduler;
 
 
 import org.apache.rocketmq.eventbridge.adapter.runtime.env.Environment;
-import org.apache.rocketmq.eventbridge.adapter.runtime.pipline.callback.TaskCallBack;
+import org.apache.rocketmq.eventbridge.adapter.runtime.pipline.callback.TaskEventCallBack;
 import org.apache.rocketmq.eventbridge.adapter.runtime.pipline.service.PipelineService;
 import org.apache.rocketmq.eventbridge.adapter.runtime.task.PipelineTask;
 
@@ -14,14 +14,14 @@ public class Pipeline implements Serializable {
 
     private transient Environment environment;
     private List<PipelineTask> pipelineTaskList;
-    private List<TaskCallBack> pipelineTaskCallbacks;
+    private List<TaskEventCallBack> pipelineTaskEventCallbacks;
     private List<PipelineService> pipelineServices;
 
     public Pipeline(Environment environment) {
         this.environment = environment;
         this.environment.addPipeline(this);
         this.pipelineTaskList = new ArrayList<>();
-        this.pipelineTaskCallbacks = new ArrayList<>();
+        this.pipelineTaskEventCallbacks = new ArrayList<>();
         this.pipelineServices = new ArrayList<>();
     }
 
@@ -29,11 +29,11 @@ public class Pipeline implements Serializable {
 
     }
 
-    public TaskCallBack submit(PipelineTask pipelineTask) {
+    public TaskEventCallBack submit(PipelineTask pipelineTask) {
         this.pipelineTaskList.add(pipelineTask);
-        TaskCallBack taskCallBack = new TaskCallBack();
-        this.pipelineTaskCallbacks.add(taskCallBack);
-        return taskCallBack;
+        TaskEventCallBack taskEventCallBack = new TaskEventCallBack();
+        this.pipelineTaskEventCallbacks.add(taskEventCallBack);
+        return taskEventCallBack;
     }
 
     public Pipeline start(PipelineService pipelineService) {
@@ -61,8 +61,8 @@ public class Pipeline implements Serializable {
         return pipelineTaskList;
     }
 
-    public List<TaskCallBack> getPipelineTaskCallbacks() {
-        return pipelineTaskCallbacks;
+    public List<TaskEventCallBack> getPipelineTaskCallbacks() {
+        return pipelineTaskEventCallbacks;
     }
 
     public List<PipelineService> getPipelineServices() {
